+++ src/data-flow/service-registration.js
@@ -3,6 +3,7 @@ import { ApiContractDescriptor } from './apis/api-contract-descriptor.js';
 import { CommandFlowDescriptor } from './commands/command-flow-descriptor.js';
 import { EventFlowDescriptor } from './events/event-flow-descriptor.js';
 import { ExecutionOverviewDescriptor } from './execution/execution-overview-descriptor.js';
+import { FutureIntegrationCapabilityRegistry } from './future/future-integration-capability-registry.js';
 import { ServiceIntegrationDescriptor } from './integrations/service-integration-descriptor.js';
 import { MessagingPatternsDescriptor } from './messaging/messaging-patterns-descriptor.js';
 import { QueryFlowDescriptor } from './queries/query-flow-descriptor.js';
@@ -18,6 +19,7 @@ export function addDataFlow(services) {
   services.registerSingleton('CommandFlowDescriptor', () => new CommandFlowDescriptor());
   services.registerSingleton('EventFlowDescriptor', () => new EventFlowDescriptor());
   services.registerSingleton('ExecutionOverviewDescriptor', () => new ExecutionOverviewDescriptor());
+  services.registerSingleton('FutureIntegrationCapabilityRegistry', () => new FutureIntegrationCapabilityRegistry());
   services.registerSingleton('MessagingPatternsDescriptor', () => new MessagingPatternsDescriptor());
   services.registerSingleton('QueryFlowDescriptor', () => new QueryFlowDescriptor());
   services.registerSingleton('ServiceIntegrationDescriptor', () => new ServiceIntegrationDescriptor());